Change double to float (great performance enhancement)

+++ src/InOneWeekend/hittable.h
@@ -19,7 +19,7 @@ struct hit_record {
     point3 p;
     vec3 normal;
     material* mat_ptr;
-    double t;
+    float t;
     bool front_face;
 
     __host__ __device__ inline void set_face_normal(const ray& r, const vec3& outward_normal) {
@@ -30,7 +30,7 @@ struct hit_record {
 
 class hittable {
    public:
-    __host__ __device__ virtual bool hit(const ray& r, double t_min, double t_max, hit_record& rec) = 0;
+    __host__ __device__ virtual bool hit(const ray& r, float t_min, float t_max, hit_record& rec) = 0;
 };
 
 #endif
