+++ src/InOneWeekend/sphere.h
@@ -20,11 +20,11 @@ class sphere : public hittable {
     public:
         __host__ __device__ sphere() {}
 
-        __host__ __device__ sphere(point3 cen, double r, material* m)
+        __host__ __device__ sphere(point3 cen, float r, material* m)
             : center(cen), radius(r), mat_ptr(m) {};
 
         __host__ __device__ bool hit(
-            const ray& r, double t_min, double t_max, hit_record& rec);
+            const ray& r, float t_min, float t_max, hit_record& rec);
         
         __host__ __device__ void whoami() {
             printf("Sphere!\n");
@@ -44,12 +44,12 @@ class sphere : public hittable {
 
     public:
         point3 center;
-        double radius;
+        float radius;
         material* mat_ptr;
 };
 
 __host__ __device__
-bool sphere::hit(const ray& r, double t_min, double t_max, hit_record& rec) {
+bool sphere::hit(const ray& r, float t_min, float t_max, hit_record& rec) {
     vec3 oc = r.origin() - center;
     auto a = r.direction().length_squared();
     auto half_b = dot(oc, r.direction());
